requirements-dev.txt
django-webpack-loader
coverage
coveralls
djangorestframework
mysqlclient
